Refactor placeholders for format to match storage
